Move do_icall into separate file
Moves do_icall into its own file - this seems to help with generating useable debug browser symbols
Enables the TEST_ICALL_SYMBOL_MAP define for testing purposes.
