+++ src/mono/mono/mini/CMakeLists.txt
@@ -239,6 +239,8 @@ set(interp_sources
     interp/interp.c
     interp/interp-intrins.h
     interp/interp-intrins.c
+    interp/interp-icalls.h
+    interp/interp-icalls.c
     interp/mintops.h
     interp/mintops.c
     interp/transform.c
